timetable_downloader/pyproject.toml
[project]
name = "timetable-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "discord-webhook>=1.4.1",
    "playwright>=1.58.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
]
